добавил отрисовку всех кнопок в header
1. оптимизировал файлы свои и Миши, удалил свой паршиал js my-library-render, всё перенёс в header-setup.js.
2. проработал все сценарии удаления/добавления в мою библиотеку фильмов из модалки.
3.Надо прикрутить стили Полины к пустому окну моей библиотеки (Полина сказала что стили для картинки и p, которая отображается при пустой библиотеке сделала(я их не увидел)
